move benchmarks to another file due to the unit test expectations and line numbers
